update hash join file

+++ Codes/hash.py
@@ -11,32 +11,21 @@ def hash_join_file(folderName,memory,pageSize):
         #vide le contenu
         delete_file("T",folderName+"_hash")
     
-    if not os.path.exists('Data/'+folderName+"_hash_temp"):
-        os.makedirs('Data/'+folderName+"_hash_temp")
-    else:
-        #vide le contenu
-        delete_file("S",folderName+"_hash_temp")
     
     nbPageR=len([f for f in os.listdir("Data/"+folderName) if ("R") in f])
     nbPageS=len([f for f in os.listdir("Data/"+folderName) if ("S") in f])
-    iPageR=1
-    tempString=""
     T=[]
     iT=1
-    while iPageR<=nbPageR:
+
+    if nbPageR<=memory-2:
+        print("a")
         H=dict()
-        tempdb=[]
-        iS=1
-        for i in range((memory-2)):
-            if iPageR<=nbPageR:
-                db=read_X_pages(folderName+"/R",iPageR,1)
-                for j in range(len(db.index)):
-                    key=hash(int(db["Y"].get(j)))
-                    assert key not in H, "Colision"
-                    H[key]=db["X"].get(j)
-                iPageR+=1
+        db=read_X_pages(folderName+"/R",1,nbPageR)
+        for i in range(len(db.index)):
+            key=hash(int(db["Y"].get(i)))
+            H[key]=db["X"].get(i)
         for i in range(nbPageS):
-            db=read_X_pages(folderName+tempString+"/S",i+1,1)
+            db=read_X_pages(folderName+"/S",i+1,1)
             for j in range(len(db.index)):
                 key=hash(int(db["Y"].get(j)))
                 if key in H:
@@ -47,23 +36,61 @@ def hash_join_file(folderName,memory,pageSize):
                         T.to_csv('Data/'+folderName+"_hash/T_"+str(iT)+".csv",sep=',',index=False)
                         iT+=1
                         T=[]
-                else:
-                    tempdb.append((db["Y"].get(j),db["Z"].get(j)))
-                    if len(tempdb)==pageSize:
-                        tempdb=pd.DataFrame(tempdb,columns=['Y','Z'])
-                        delete_file("S_"+str(iS)+".csv",folderName+"_hash_temp")
-                        tempdb.to_csv('Data/'+folderName+"_hash_temp/S_"+str(iS)+".csv",sep=',',index=False)
-                        iS+=1
-                        tempdb=[]
-        if tempdb:
-            tempdb=pd.DataFrame(tempdb,columns=['Y','Z'])
-            tempdb.to_csv('Data/'+folderName+"_hash_temp/S_"+str(iS)+".csv",sep=',',index=False)
-            iS+=1
-        tempString="_hash_temp"
-        nbPageS=iS-1
-    if T:
-        T=pd.DataFrame(T,columns=['X','Y','Z'])
-        T.to_csv('Data/'+folderName+"_hash/T_"+str(iT)+".csv",sep=',',index=False)
+        if T:
+            T=pd.DataFrame(T,columns=['X','Y','Z'])
+            T.to_csv('Data/'+folderName+"_hash/T_"+str(iT)+".csv",sep=',',index=False)
+
+    else:
+
+        if not os.path.exists('Data/'+folderName+"_hash_temp"):
+            os.makedirs('Data/'+folderName+"_hash_temp")
+        else:
+            #vide le contenu
+            delete_file("S",folderName+"_hash_temp")
+
+        iPageR=1
+        tempString=""
+        while iPageR<=nbPageR:
+            H=dict()
+            tempdb=[]
+            iS=1
+            for i in range((memory-2)):
+                if iPageR<=nbPageR:
+                    db=read_X_pages(folderName+"/R",iPageR,1)
+                    for j in range(len(db.index)):
+                        key=hash(int(db["Y"].get(j)))
+                        assert key not in H, "Colision"
+                        H[key]=db["X"].get(j)
+                    iPageR+=1
+            for i in range(nbPageS):
+                db=read_X_pages(folderName+tempString+"/S",i+1,1)
+                for j in range(len(db.index)):
+                    key=hash(int(db["Y"].get(j)))
+                    if key in H:
+                        T.append((H[key],db["Y"].get(j),db["Z"].get(j)))
+                        if len(T)==pageSize:
+                            print(T)
+                            T=pd.DataFrame(T,columns=['X','Y','Z'])
+                            T.to_csv('Data/'+folderName+"_hash/T_"+str(iT)+".csv",sep=',',index=False)
+                            iT+=1
+                            T=[]
+                    else:
+                        tempdb.append((db["Y"].get(j),db["Z"].get(j)))
+                        if len(tempdb)==pageSize:
+                            tempdb=pd.DataFrame(tempdb,columns=['Y','Z'])
+                            delete_file("S_"+str(iS)+".csv",folderName+"_hash_temp")
+                            tempdb.to_csv('Data/'+folderName+"_hash_temp/S_"+str(iS)+".csv",sep=',',index=False)
+                            iS+=1
+                            tempdb=[]
+            if tempdb:
+                tempdb=pd.DataFrame(tempdb,columns=['Y','Z'])
+                tempdb.to_csv('Data/'+folderName+"_hash_temp/S_"+str(iS)+".csv",sep=',',index=False)
+                iS+=1
+            tempString="_hash_temp"
+            nbPageS=iS-1
+        if T:
+            T=pd.DataFrame(T,columns=['X','Y','Z'])
+            T.to_csv('Data/'+folderName+"_hash/T_"+str(iT)+".csv",sep=',',index=False)
